+++ src/mizdb_inlines/views.py
@@ -1,3 +1,6 @@
+import operator
+from functools import reduce
+
 from django.forms import all_valid
 from django.views.generic.edit import ModelFormMixin
 
@@ -27,6 +30,7 @@ class InlineFormsetMixin(ModelFormMixin):
     def get_context_data(self, formsets=None, **kwargs):
         ctx = super().get_context_data(**kwargs)
         ctx["formsets"] = formsets or self.get_formsets(parent_instance=ctx["form"].instance)
+        ctx["formset_media"] = reduce(operator.add, (fs.media for fs in ctx["formsets"]))
         return ctx
 
     def formsets_valid(self, formsets):