+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [unreleased]
+
+### added
+- the combined media of all formsets is now available to the template context under the name `formset_media` 
+
 ## 0.1.0 (2023-08-31)
 
 - initial release